requirements.txt
asgiref==3.7.2
Django==4.2.3
django-getenv==1.3.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.6
gunicorn==20.1.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.5.0
django-cors-headers==4.2.0

